+++ www/go/modules/community/addressbook/model/Contact.php
@@ -976,6 +976,10 @@ class Contact extends AclItemEntity {
 			}
 		}
 
+		if(!empty($this->notes)) {
+			$keywords[] = $this->notes;
+		}
+
 		return $keywords;
 	}